Moved member NQL serialization into the shared filter layer
ref https://linear.app/ghost/issue/BER-3414
Comments already serialized through shared filter modules, so leaving member NQL inside the hook kept the abstraction split and harder to reason about.
