+++ apps/posts/src/views/members/hooks/use-members-filter-state.ts
@@ -1,14 +1,15 @@
 import {useMemo} from 'react';
 import {useBrowseSettings} from '@tryghost/admin-x-framework/api/settings';
 import type {Filter} from '@tryghost/shade';
-import {canonicalizeFilter} from '@src/views/filters/canonical-filter';
 import {deriveFilterFlags} from '@src/views/filters/filter-flags';
+import {buildMemberNqlFilter} from '@src/views/filters/member-nql';
 import {parsePredicateParams, serializePredicateParams} from '@src/views/filters/url-predicate-params';
 import {UrlFilterStateOptions, useUrlFilterState} from '@src/views/filters/use-url-filter-state';
 import {getSiteTimezone} from '@src/utils/get-site-timezone';
 import {deriveMemberFilterMetadata, MemberFilterColumnMetadata} from './member-filter-metadata';
-import moment from 'moment-timezone';
 import nql from '@tryghost/nql-lang';
+
+export {buildMemberNqlFilter};
 
 /**
  * Member filter field keys - single source of truth for filter definitions
@@ -48,235 +49,6 @@ export type MemberFilterField = typeof MEMBER_FILTER_FIELDS[number];
 
 // Fields that support multiselect (comma-separated values in URL)
 const MULTISELECT_FIELDS = new Set<string>(['label', 'offer_redemptions']);
-
-/**
- * Escape a string for NQL (escape single quotes)
- */
-function escapeNqlString(value: string): string {
-    return '\'' + value.replace(/'/g, '\\\'') + '\'';
-}
-
-function formatDateFilterValue(value: string, relation: string, timezone: string): string {
-    const trimmedValue = value.trim();
-    const localDate = moment.tz(trimmedValue, 'YYYY-MM-DD', true, timezone);
-
-    if (!localDate.isValid()) {
-        return trimmedValue;
-    }
-
-    const boundaryDate = (relation === '>' || relation === '<=')
-        ? localDate.set({hour: 23, minute: 59, second: 59})
-        : localDate.set({hour: 0, minute: 0, second: 0});
-
-    return boundaryDate.utc().format('YYYY-MM-DD HH:mm:ss');
-}
-
-function normalizeNewsletterSubscriptionValue(value: string): 'true' | 'false' | string {
-    if (value === 'subscribed') {
-        return 'true';
-    }
-
-    if (value === 'unsubscribed') {
-        return 'false';
-    }
-
-    return value;
-}
-
-function isSubscribedNewsletterFilter(operator: string, value: string): boolean {
-    const normalizedValue = normalizeNewsletterSubscriptionValue(value);
-    const isInverseOperator = operator === 'is-not' || operator === 'is_not';
-
-    return (operator === 'is' && normalizedValue === 'true') || (isInverseOperator && normalizedValue === 'false');
-}
-
-/**
- * Map UI operator names to NQL operators
- */
-function getFilterRelationOperator(relation: string): string {
-    const relationMap: Record<string, string> = {
-        'is-less': '<',
-        'is-or-less': '<=',
-        is: '',
-        'is-not': '-',
-        'is-greater': '>',
-        'is-or-greater': '>=',
-        contains: '~',
-        'does-not-contain': '-~',
-        not_contains: '-~',
-        'starts-with': '~^',
-        'ends-with': '~$',
-        // Shade filter operators (mapped to our internal names)
-        before: '<',
-        after: '>',
-        is_not: '-',
-        is_any_of: '',
-        is_not_any_of: '-',
-        greater_than: '>',
-        less_than: '<',
-        equals: '',
-        not_equals: '-'
-    };
-
-    return relationMap[relation] ?? '';
-}
-
-/**
- * Build NQL filter string from filter array
- */
-export function buildMemberNqlFilter(filters: Filter[], options: {timezone?: string} = {}): string | undefined {
-    const parts: string[] = [];
-    const timezone = options.timezone ?? 'Etc/UTC';
-
-    for (const filter of filters) {
-        if (!filter.values[0] && filter.values[0] !== 0) {
-            continue;
-        }
-
-        const field = filter.field;
-        const operator = filter.operator;
-        const value = filter.values[0];
-        const relationStr = getFilterRelationOperator(operator);
-
-        // Handle newsletter filters with dynamic keys (e.g., newsletters.weekly-digest)
-        if (field.startsWith('newsletters.')) {
-            const slug = field.replace('newsletters.', '');
-            const subscriptionStatus = String(value);
-
-            if (isSubscribedNewsletterFilter(operator, subscriptionStatus)) {
-                parts.push(`(newsletters.slug:${slug}+email_disabled:0)`);
-            } else {
-                parts.push(`(newsletters.slug:-${slug},email_disabled:1)`);
-            }
-            continue;
-        }
-
-        switch (field) {
-        // Text filters (name, email)
-        case 'name':
-        case 'email': {
-            const escapedValue = escapeNqlString(String(value));
-            parts.push(`${field}:${relationStr}${escapedValue}`);
-            break;
-        }
-
-        // Array filters (label, tier_id, offer_redemptions)
-        case 'label':
-        case 'tier_id':
-        case 'offer_redemptions': {
-            if (Array.isArray(filter.values) && filter.values.length > 0) {
-                const filterValue = '[' + filter.values.join(',') + ']';
-                parts.push(`${field}:${relationStr}${filterValue}`);
-            } else if (value) {
-                parts.push(`${field}:${relationStr}${value}`);
-            }
-            break;
-        }
-
-        // Status filter
-        case 'status': {
-            parts.push(`status:${relationStr}${value}`);
-            break;
-        }
-
-        // Subscribed filter (complex NQL)
-        case 'subscribed': {
-            if (value === 'email-disabled') {
-                if (operator === 'is') {
-                    parts.push('(email_disabled:1)');
-                } else {
-                    parts.push('(email_disabled:0)');
-                }
-            } else if (operator === 'is' || operator === 'is_any_of') {
-                if (value === 'subscribed') {
-                    parts.push('(subscribed:true+email_disabled:0)');
-                } else {
-                    parts.push('(subscribed:false+email_disabled:0)');
-                }
-            } else {
-                // is-not
-                if (value === 'subscribed') {
-                    parts.push('(subscribed:false,email_disabled:1)');
-                } else {
-                    parts.push('(subscribed:true,email_disabled:1)');
-                }
-            }
-            break;
-        }
-
-        // Newsletter filters (dynamic)
-        case 'newsletters': {
-            // Value format: "slug:subscribed" or "slug:unsubscribed"
-            const [slug, subscriptionStatus] = String(value).split(':');
-
-            if (isSubscribedNewsletterFilter(operator, subscriptionStatus)) {
-                parts.push(`(newsletters.slug:${slug}+email_disabled:0)`);
-            } else {
-                parts.push(`(newsletters.slug:-${slug},email_disabled:1)`);
-            }
-            break;
-        }
-
-        // Date filters
-        case 'last_seen_at':
-        case 'created_at':
-        case 'subscriptions.start_date':
-        case 'subscriptions.current_period_end': {
-            const dateValue = formatDateFilterValue(String(value), relationStr, timezone);
-            parts.push(`${field}:${relationStr}'${dateValue}'`);
-            break;
-        }
-
-        // Number filters
-        case 'email_count':
-        case 'email_opened_count':
-        case 'email_open_rate': {
-            parts.push(`${field}:${relationStr}${value}`);
-            break;
-        }
-
-        // Subscription plan interval
-        case 'subscriptions.plan_interval': {
-            parts.push(`subscriptions.plan_interval:${relationStr}${value}`);
-            break;
-        }
-
-        // Subscription status
-        case 'subscriptions.status': {
-            parts.push(`subscriptions.status:${relationStr}${value}`);
-            break;
-        }
-
-        // Resource filters — value is a post/page ID
-        case 'signup':
-        case 'conversion':
-        case 'emails.post_id':
-        case 'opened_emails.post_id':
-        case 'clicked_links.post_id': {
-            parts.push(`${field}:${relationStr}'${value}'`);
-            break;
-        }
-
-        // Audience feedback — operator is the score, value is the post ID
-        case 'newsletter_feedback': {
-            // NQL: (feedback.post_id:'<postId>'+feedback.score:<0or1>)
-            const score = operator; // '1' = More like this, '0' = Less like this
-            parts.push(`(feedback.post_id:'${value}'+feedback.score:${score})`);
-            break;
-        }
-
-        default:
-            // Generic fallback
-            if (typeof value === 'string' && value.includes(' ')) {
-                parts.push(`${field}:${relationStr}'${value}'`);
-            } else {
-                parts.push(`${field}:${relationStr}${value}`);
-            }
-        }
-    }
-
-    return canonicalizeFilter(parts);
-}
 
 function parseLegacySubscribedFilter(filterNode: Record<string, unknown>): Filter | undefined {
     const comparator = ('$and' in filterNode && Array.isArray(filterNode.$and))
